+++ symbolic-ppdb/src/format/metadata.rs
@@ -2,9 +2,10 @@ use std::convert::TryInto;
 use std::fmt;
 use std::ops::{Index, IndexMut};
 
+use symbolic_common::Uuid;
 use watto::Pod;
 
-use super::{FormatError, FormatErrorKind};
+use super::{FormatError, FormatErrorKind, PortablePdb};
 
 /// An enumeration of all table types in ECMA-335 and Portable PDB.
 #[repr(usize)]
@@ -60,9 +61,63 @@ pub enum TableType {
     DummyEmpty = 0x3F,
 }
 
+impl From<usize> for TableType {
+    fn from(value: usize) -> Self {
+        match value {
+            x if x == Self::Assembly as usize => Self::Assembly,
+            x if x == Self::AssemblyProcessor as usize => Self::AssemblyProcessor,
+            x if x == Self::AssemblyRef as usize => Self::AssemblyRef,
+            x if x == Self::AssemblyRefOs as usize => Self::AssemblyRefOs,
+            x if x == Self::AssemblyRefProcessor as usize => Self::AssemblyRefProcessor,
+            x if x == Self::ClassLayout as usize => Self::ClassLayout,
+            x if x == Self::Constant as usize => Self::Constant,
+            x if x == Self::CustomAttribute as usize => Self::CustomAttribute,
+            x if x == Self::DeclSecurity as usize => Self::DeclSecurity,
+            x if x == Self::EventMap as usize => Self::EventMap,
+            x if x == Self::Event as usize => Self::Event,
+            x if x == Self::ExportedType as usize => Self::ExportedType,
+            x if x == Self::Field as usize => Self::Field,
+            x if x == Self::FieldLayout as usize => Self::FieldLayout,
+            x if x == Self::FieldMarshal as usize => Self::FieldMarshal,
+            x if x == Self::FieldRVA as usize => Self::FieldRVA,
+            x if x == Self::File as usize => Self::File,
+            x if x == Self::GenericParam as usize => Self::GenericParam,
+            x if x == Self::GenericParamConstraint as usize => Self::GenericParamConstraint,
+            x if x == Self::ImplMap as usize => Self::ImplMap,
+            x if x == Self::InterfaceImpl as usize => Self::InterfaceImpl,
+            x if x == Self::ManifestResource as usize => Self::ManifestResource,
+            x if x == Self::MemberRef as usize => Self::MemberRef,
+            x if x == Self::MethodDef as usize => Self::MethodDef,
+            x if x == Self::MethodImpl as usize => Self::MethodImpl,
+            x if x == Self::MethodSemantics as usize => Self::MethodSemantics,
+            x if x == Self::MethodSpec as usize => Self::MethodSpec,
+            x if x == Self::Module as usize => Self::Module,
+            x if x == Self::ModuleRef as usize => Self::ModuleRef,
+            x if x == Self::NestedClass as usize => Self::NestedClass,
+            x if x == Self::Param as usize => Self::Param,
+            x if x == Self::Property as usize => Self::Property,
+            x if x == Self::PropertyMap as usize => Self::PropertyMap,
+            x if x == Self::StandAloneSig as usize => Self::StandAloneSig,
+            x if x == Self::TypeDef as usize => Self::TypeDef,
+            x if x == Self::TypeRef as usize => Self::TypeRef,
+            x if x == Self::TypeSpec as usize => Self::TypeSpec,
+            x if x == Self::CustomDebugInformation as usize => Self::CustomDebugInformation,
+            x if x == Self::Document as usize => Self::Document,
+            x if x == Self::ImportScope as usize => Self::ImportScope,
+            x if x == Self::LocalConstant as usize => Self::LocalConstant,
+            x if x == Self::LocalScope as usize => Self::LocalScope,
+            x if x == Self::LocalVariable as usize => Self::LocalVariable,
+            x if x == Self::MethodDebugInformation as usize => Self::MethodDebugInformation,
+            x if x == Self::StateMachineMethod as usize => Self::StateMachineMethod,
+            _ => Self::DummyEmpty,
+        }
+    }
+}
+
 /// A table in a Portable PDB file.
-#[derive(Default, Clone, Copy)]
+#[derive(Clone, Copy)]
 pub struct Table<'data> {
+    type_: TableType,
     /// The number of rows in the table.
     pub rows: usize,
     /// The width in bytes of one table row.
@@ -162,9 +217,46 @@ impl<'data> Table<'data> {
     /// Returns the the bytes of the `idx`th row, if any.
     ///
     /// Note that table row indices are 1-based!
-    fn get_row(&self, idx: usize) -> Option<&'data [u8]> {
+    pub(crate) fn get_row(&self, idx: usize) -> Result<Row, FormatError> {
         idx.checked_sub(1)
             .and_then(|idx| self.contents.get(idx * self.width..(idx + 1) * self.width))
+            .map(|data| Row { data, table: self })
+            .ok_or_else(|| FormatErrorKind::RowIndexOutOfBounds(self.type_, idx).into())
+    }
+}
+
+/// A row in a [Table].
+#[derive(Debug, Clone, Copy)]
+pub(crate) struct Row<'data> {
+    data: &'data [u8],
+    table: &'data Table<'data>,
+}
+
+impl<'data> Row<'data> {
+    /// Reads the `col` cell in the given table as a `u32`.
+    ///
+    /// This returns an error if the indices are out of bounds for the table
+    /// or the cell is too wide for a `u32`.
+    ///
+    /// Note that row and column indices are 1-based!
+    pub(crate) fn get_col_u32(&self, col: usize) -> Result<u32, FormatError> {
+        if !(1..=6).contains(&col) {
+            return Err(FormatErrorKind::ColIndexOutOfBounds(self.table.type_, col).into());
+        }
+        let Column { offset, width } = self.table.columns[col - 1];
+        match width {
+            1 => Ok(self.data[offset] as u32),
+            2 => {
+                let bytes = &self.data[offset..offset + 2];
+                Ok(u16::from_ne_bytes(bytes.try_into().unwrap()) as u32)
+            }
+            4 => {
+                let bytes = &self.data[offset..offset + 4];
+                Ok(u32::from_ne_bytes(bytes.try_into().unwrap()))
+            }
+
+            _ => Err(FormatErrorKind::ColumnWidth(self.table.type_, col, width).into()),
+        }
     }
 }
 
@@ -255,8 +347,13 @@ impl<'data> MetadataStream<'data> {
 
         // TODO: verify major/minor version
         // TODO: verify reserved
-
-        let mut tables = [Table::default(); 64];
+        let mut tables = [Table {
+            type_: TableType::DummyEmpty,
+            rows: usize::default(),
+            width: usize::default(),
+            columns: [Column::default(); 6],
+            contents: <&[u8]>::default(),
+        }; 64];
         for (i, table) in tables.iter_mut().enumerate() {
             if (header.valid_tables >> i & 1) == 0 {
                 continue;
@@ -264,7 +361,7 @@ impl<'data> MetadataStream<'data> {
 
             let (len, rest_) = u32::ref_from_prefix(rest).ok_or(FormatErrorKind::InvalidLength)?;
             rest = rest_;
-
+            table.type_ = TableType::from(i);
             table.rows = *len as usize;
         }
 
@@ -287,47 +384,6 @@ impl<'data> MetadataStream<'data> {
         result.set_contents(table_contents);
 
         Ok(result)
-    }
-
-    /// Returns the bytes of the `idx`th row of the `table` table, if any.
-    ///
-    /// Note that table row indices are 1-based!
-    fn get_row(&self, table: TableType, idx: usize) -> Option<&'data [u8]> {
-        self[table].get_row(idx)
-    }
-
-    /// Reads the `(row, col)` cell in the given table as a `u32`.
-    ///
-    /// This returns an error if the indices are out of bounds for the table
-    /// or the cell is too wide for a `u32`.
-    ///
-    /// Note that row and column indices are 1-based!
-    pub(crate) fn get_table_cell_u32(
-        &self,
-        table: TableType,
-        row: usize,
-        col: usize,
-    ) -> Result<u32, FormatError> {
-        let row = self
-            .get_row(table, row)
-            .ok_or(FormatErrorKind::RowIndexOutOfBounds(table, row))?;
-        if !(1..=6).contains(&col) {
-            return Err(FormatErrorKind::ColIndexOutOfBounds(table, col).into());
-        }
-        let Column { offset, width } = self[table].columns[col - 1];
-        match width {
-            1 => Ok(row[offset] as u32),
-            2 => {
-                let bytes = &row[offset..offset + 2];
-                Ok(u16::from_ne_bytes(bytes.try_into().unwrap()) as u32)
-            }
-            4 => {
-                let bytes = &row[offset..offset + 4];
-                Ok(u32::from_ne_bytes(bytes.try_into().unwrap()))
-            }
-
-            _ => Err(FormatErrorKind::ColumnWidth(table, col, width).into()),
-        }
     }
 
     /// Sets the column widths of all tables in this stream.
@@ -809,5 +865,148 @@ impl<'data> Index<TableType> for MetadataStream<'data> {
 impl<'data> IndexMut<TableType> for MetadataStream<'data> {
     fn index_mut(&mut self, index: TableType) -> &mut Self::Output {
         &mut self.tables[index as usize]
+    }
+}
+
+/// An iterator over CustomDebugInformation of a specific Kind.
+/// See [CustomDebugInformation](https://github.com/dotnet/runtime/blob/main/docs/design/specs/PortablePdb-Metadata.md#customdebuginformation-table-0x37).
+#[derive(Debug, Clone)]
+pub(crate) struct CustomDebugInformationIterator<'data> {
+    table: Table<'data>,
+    /// Which kind of CustomDebugInformation we want to filter.
+    /// We only store the offset in the GUID table to avoid lookups every time.
+    kind: Option<u32>,
+    /// Current row in the whole table (not just the filtered kind).
+    /// Note that the row is 1-based, to align with the rest of the crate APIs.
+    row: usize,
+}
+
+impl<'data> CustomDebugInformationIterator<'data> {
+    pub(crate) fn new(ppdb: &PortablePdb<'data>, filter_kind: Uuid) -> Result<Self, FormatError> {
+        let md_stream = ppdb
+            .metadata_stream
+            .as_ref()
+            .ok_or(FormatErrorKind::NoMetadataStream)?;
+
+        let kind = ppdb
+            .guid_stream
+            .as_ref()
+            .ok_or(FormatErrorKind::NoGuidStream)?
+            .get_offset(filter_kind);
+
+        Ok(CustomDebugInformationIterator {
+            table: md_stream[TableType::CustomDebugInformation],
+            kind,
+            row: 1,
+        })
+    }
+}
+
+macro_rules! ok_or_return {
+    ( $a:expr ) => {
+        match $a {
+            Ok(value) => value,
+            Err(err) => return Some(Err(err)),
+        }
+    };
+}
+
+impl<'data> Iterator for CustomDebugInformationIterator<'data> {
+    type Item = Result<CustomDebugInformation, FormatError>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let expected_kind_offset = self.kind?;
+        // Find the first row in the table matching the desired Kind.
+        while self.row <= self.table.rows {
+            let row = ok_or_return!(self.table.get_row(self.row));
+            self.row += 1;
+
+            let kind_offset = ok_or_return!(row.get_col_u32(2));
+
+            if kind_offset == expected_kind_offset {
+                // Column 1 contains a Parent coded with HasCustomDebugInformation on the lower 5 bits
+                let parent = ok_or_return!(row.get_col_u32(1));
+                let value = parent >> 5;
+                let tag = ok_or_return!(CustomDebugInformationTag::from(parent & 0b11111));
+
+                let blob = ok_or_return!(row.get_col_u32(3));
+                return Some(Ok(CustomDebugInformation { tag, value, blob }));
+            }
+        }
+        None
+    }
+}
+
+#[derive(Debug, Clone, Copy)]
+pub(crate) struct CustomDebugInformation {
+    pub(crate) tag: CustomDebugInformationTag,
+    pub(crate) value: u32,
+    pub(crate) blob: u32,
+}
+
+/// See [CustomDebugInformation](https://github.com/dotnet/runtime/blob/main/docs/design/specs/PortablePdb-Metadata.md#customdebuginformation-table-0x37).
+#[derive(Debug, Clone, Copy)]
+pub(crate) enum CustomDebugInformationTag {
+    MethodDef = 0,
+    Field = 1,
+    TypeRef = 2,
+    TypeDef = 3,
+    Param = 4,
+    InterfaceImpl = 5,
+    MemberRef = 6,
+    Module = 7,
+    DeclSecurity = 8,
+    Property = 9,
+    Event = 10,
+    StandAloneSig = 11,
+    ModuleRef = 12,
+    TypeSpec = 13,
+    Assembly = 14,
+    AssemblyRef = 15,
+    File = 16,
+    ExportedType = 17,
+    ManifestResource = 18,
+    GenericParam = 19,
+    GenericParamConstraint = 20,
+    MethodSpec = 21,
+    Document = 22,
+    LocalScope = 23,
+    LocalVariable = 24,
+    LocalConstant = 25,
+    ImportScope = 26,
+}
+
+impl CustomDebugInformationTag {
+    fn from(value: u32) -> Result<Self, FormatError> {
+        Ok(match value {
+            x if x == Self::MethodDef as u32 => Self::MethodDef,
+            x if x == Self::Field as u32 => Self::Field,
+            x if x == Self::TypeRef as u32 => Self::TypeRef,
+            x if x == Self::TypeDef as u32 => Self::TypeDef,
+            x if x == Self::Param as u32 => Self::Param,
+            x if x == Self::InterfaceImpl as u32 => Self::InterfaceImpl,
+            x if x == Self::MemberRef as u32 => Self::MemberRef,
+            x if x == Self::Module as u32 => Self::Module,
+            x if x == Self::DeclSecurity as u32 => Self::DeclSecurity,
+            x if x == Self::Property as u32 => Self::Property,
+            x if x == Self::Event as u32 => Self::Event,
+            x if x == Self::StandAloneSig as u32 => Self::StandAloneSig,
+            x if x == Self::ModuleRef as u32 => Self::ModuleRef,
+            x if x == Self::TypeSpec as u32 => Self::TypeSpec,
+            x if x == Self::Assembly as u32 => Self::Assembly,
+            x if x == Self::AssemblyRef as u32 => Self::AssemblyRef,
+            x if x == Self::File as u32 => Self::File,
+            x if x == Self::ExportedType as u32 => Self::ExportedType,
+            x if x == Self::ManifestResource as u32 => Self::ManifestResource,
+            x if x == Self::GenericParam as u32 => Self::GenericParam,
+            x if x == Self::GenericParamConstraint as u32 => Self::GenericParamConstraint,
+            x if x == Self::MethodSpec as u32 => Self::MethodSpec,
+            x if x == Self::Document as u32 => Self::Document,
+            x if x == Self::LocalScope as u32 => Self::LocalScope,
+            x if x == Self::LocalVariable as u32 => Self::LocalVariable,
+            x if x == Self::LocalConstant as u32 => Self::LocalConstant,
+            x if x == Self::ImportScope as u32 => Self::ImportScope,
+            _ => return Err(FormatErrorKind::InvalidCustomDebugInformationTag(value).into()),
+        })
     }
 }