+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Features**:
 
 - `PortablePdbDebugSession` now returns files referenced in the Portable PDB file. ([#729](https://github.com/getsentry/symbolic/pull/729))
+- `PortablePdbDebugSession` now returns source files embedded in the Portable PDB file. ([#734](https://github.com/getsentry/symbolic/pull/734))
 
 **Breaking changes**: